Separate celebration script localization
