+++ assets/js/web-components/prpl-suggested-task.js
@@ -508,6 +508,7 @@ customElements.define(
 								new CustomEvent( event, {
 									detail: {
 										element: el,
+										taskList,
 									},
 								} )
 							);
